# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FPS Brandpreventie - Platform voor brandpreventieve gebouwvoorzieningen
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/wachtwoord-wijzigen:
    post:
      operationId: wachtwoordWijzigen
      tags: [auth]
      summary: Eigen wachtwoord wijzigen (ingelogd)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WachtwoordWijzigen"
      responses:
        "204":
          description: Wachtwoord gewijzigd
  /auth/taal:
    post:
      operationId: taalWijzigen
      tags: [auth]
      summary: Eigen taalvoorkeur wijzigen (ingelogd)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaalWijzigen"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthGebruiker"

  # ── ABONNEMENTEN ─────────────────────────────────────────────────────────
  /abonnementen:
    get:
      operationId: listAbonnementen
      tags: [abonnementen]
      summary: Lijst van abonnementen
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Abonnement"
    post:
      operationId: createAbonnement
      tags: [abonnementen]
      summary: Abonnement aanmaken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbonnementInput"
      responses:
        "201":
          description: Aangemaakt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Abonnement"
  /abonnementen/{id}:
    get:
      operationId: getAbonnement
      tags: [abonnementen]
      summary: Abonnementdetails ophalen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Abonnement"
    patch:
      operationId: updateAbonnement
      tags: [abonnementen]
      summary: Abonnement bijwerken
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbonnementUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Abonnement"

  # ── AUTHENTICATIE ────────────────────────────────────────────────────────
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Inloggen met e-mail en wachtwoord (stap 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResultaat"
...
  /auth/2fa/activeren:
    post:
      operationId: tweeFactorActiveren
      tags: [auth]
      summary: Activeer tweestapsverificatie met de eerste code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthGebruiker"
  /auth/2fa/verify:
    post:
      operationId: tweeFactorVerify
      tags: [auth]
      summary: Verifieer tweestapsverificatie-code (stap 2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthGebruiker"
...
components:
  schemas:
...
    WachtwoordWijzigen:
      type: object
      required: [huidig_wachtwoord, nieuw_wachtwoord]
      properties:
        huidig_wachtwoord:
          type: string
        nieuw_wachtwoord:
          type: string
    TaalWijzigen:
      type: object
      required: [taal]
      properties:
        taal:
          type: string
          enum: [nl, en, de, fr, ar, tr]
...
    LoginInput:
      type: object
      required: [email, wachtwoord]
      properties:
        email:
          type: string
        wachtwoord:
          type: string
    LoginResultaat:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [setup_2fa, verify_2fa]
...
    CodeInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    AuthGebruiker:
      type: object
      required: [id, naam, email, rol, bevoegdheden]
      properties:
        id:
          type: integer
        naam:
          type: string
        email:
          type: string
        rol:
          type: string
          enum: [hoofdbeheerder, beheerder, monteur, controleur, klant, gebruiker]
        avatar_url:
          type: ["string", "null"]
        bedrijfskleuren:
          type: ["string", "null"]
        taal:
          type: string
          enum: [nl, en, de, fr, ar, tr]
        nieuw_apparaat:
          type: boolean
        nieuw_ip:
          type: boolean
        functietitels:
          type: array
          items:
            type: string
        bevoegdheden:
          type: object
          additionalProperties:
            type: integer

    # ── ABONNEMENT ───────────────────────────────────────────────────────
    Abonnement:
      type: object
      required: [id, naam, niveau, prijs_per_maand, actief, aangemaakt_op]
      properties:
        id:
          type: integer
        naam:
          type: string
        niveau:
          type: string
          enum: [basis, beheer, volledig]
        prijs_per_maand:
          type: number
        max_gebouwen:
          type: ["integer", "null"]
        max_gebruikers:
          type: ["integer", "null"]
        functies:
          type: array
          items:
            type: string
        klant_naam:
          type: ["string", "null"]
        klant_email:
          type: ["string", "null"]
        start_datum:
          type: ["string", "null"]
        eind_datum:
          type: ["string", "null"]
        actief:
          type: boolean
        aangemaakt_op:
          type: string
    AbonnementInput:
      type: object
      required: [naam, niveau, prijs_per_maand]
      properties:
        naam:
          type: string
        niveau:
          type: string
        prijs_per_maand:
          type: number
        max_gebouwen:
          type: integer
        max_gebruikers:
          type: integer
        functies:
          type: array
          items:
            type: string
        klant_naam:
          type: string
        klant_email:
          type: string
        start_datum:
          type: string
        eind_datum:
          type: string
    AbonnementUpdate:
      type: object
      properties:
        naam:
          type: string
        niveau:
          type: string
        prijs_per_maand:
          type: number
        max_gebouwen:
          type: integer
        max_gebruikers:
          type: integer
        functies:
          type: array
          items:
            type: string
        klant_naam:
          type: string
        klant_email:
          type: string
        start_datum:
          type: string
        eind_datum:
          type: string
        actief:
          type: boolean
...