# lib/api-zod/src/generated/types/authGebruiker.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FPS Brandpreventie - Platform voor brandpreventieve gebouwvoorzieningen
 * OpenAPI spec version: 0.1.0
 */
import type { AuthGebruikerBevoegdheden } from './authGebruikerBevoegdheden';
import type { AuthGebruikerRol } from './authGebruikerRol';
import type { AuthGebruikerTaal } from './authGebruikerTaal';

export interface AuthGebruiker {
  id: number;
  naam: string;
  email: string;
  rol: AuthGebruikerRol;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  bedrijfskleuren?: string | null;
  taal?: AuthGebruikerTaal;
  nieuw_apparaat?: boolean;
  nieuw_ip?: boolean;
  functietitels?: string[];
  bevoegdheden: AuthGebruikerBevoegdheden;
}